refactor code to avoid circular import, still not verified working
